+++ Draggable.php
@@ -31,6 +31,16 @@ use yii\helpers\Html;
 class Draggable extends Widget
 {
     /**
+     * @inheritdoc
+     */
+    protected $clientEventMap = [
+        'create' => 'dragcreate',
+        'drag' => 'drag',
+        'stop' => 'dragstop',
+        'start' => 'dragstart',
+    ];
+
+    /**
      * Initializes the widget.
      */
     public function init()