+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 jui extension Change Log
 2.0.6 under development
 -----------------------
 
-- no changes in this release.
+- Bug #36: `yii\jui\Draggable` was using wrong event names (samdark)
 
 
 2.0.5 March 17, 2016